move sampling to separate file
