+++ tests_cpp/eigen_2d_swe_slip_wall_implicit_mixed_schwarz/gen_sample_mesh.py
@@ -1,5 +1,5 @@
 
-from pdas.prom_utils import gen_sample_mesh
+from pdas.samp_utils import gen_sample_mesh
 
 gen_sample_mesh(
     "random",
