+++ tests_cpp/parallel/eigen_2d_swe_slip_wall_implicit_hproms_schwarz_parallel/gen_sample_mesh.py
@@ -1,7 +1,7 @@
 import os
 from argparse import ArgumentParser
 
-from pdas.prom_utils import gen_sample_mesh
+from pdas.samp_utils import gen_sample_mesh
 
 parser = ArgumentParser()
 parser.add_argument("--outdir", dest="outdir")
